# mjfast-admin/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: 企业快速开发框架在线API
basePath: /admin
schemes:
- http
paths:
...
  /sys/dept/info/{deptId}:
    get:
      tags:
      - 部门管理
      summary: 获取部门信息
      produces:
      - application/json
      parameters:
      - name: deptId
        description: 部门ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回部门信息
          schema:
            $ref: '#/definitions/SysDeptEntity'
  /sys/dept/save:
    post:
      tags:
      - 部门管理
      summary: 添加部门
      produces:
      - application/json
      parameters:
      - name: body
        description: 部门对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysDeptEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/update:
    post:
      tags:
      - 部门管理
      summary: 修改部门
      produces:
      - application/json
      parameters:
      - name: body
        description: 部门对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysDeptEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/delete:
    post:
      tags:
      - 部门管理
      summary: 删除部门
      produces:
      - application/json
      parameters:
      - name: body
        description: 部门ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
      - 菜单管理
      summary: 获取菜单信息
      produces:
      - application/json
      parameters:
      - name: menuId
        description: 菜单ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/menu/save:
    post:
      tags:
      - 菜单管理
      summary: 添加菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
      - 菜单管理
      summary: 修改菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete:
    post:
      tags:
      - 菜单管理
      summary: 删除菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/log/list:
    get:
      tags:
      - 系统日志
      summary: 日志列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: key
        description: 用户名或用户操作
        in: query
        type: string
      responses:
        '200':
          description: 返回日志列表
          schema:
            $ref: '#/definitions/SysLogEntity'
  /sys/config/list:
    get:
      tags:
      - 参数管理
      summary: 参数列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: key
        description: 参数名
        in: query
        type: string
      responses:
        '200':
          description: 返回参数列表
          schema:
            $ref: '#/definitions/SysConfigEntity'
  /sys/config/info/{id}:
    get:
      tags:
      - 参数管理
      summary: 获取参数信息
      produces:
      - application/json
      parameters:
      - name: id
        description: 参数ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回参数信息
          schema:
            $ref: '#/definitions/SysConfigEntity'
  /sys/config/save:
    post:
      tags:
      - 参数管理
      summary: 添加参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysConfigEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/update:
    post:
      tags:
      - 参数管理
      summary: 修改参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysConfigEntity'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/delete:
    post:
      tags:
      - 参数管理
      summary: 删除参数
      produces:
      - application/json
      parameters:
      - name: body
        description: 参数ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/schedule/list:
    get:
      tags:
      - 定时任务
      summary: 定时任务列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: beanName
        description: spring bean名称
        in: query
        type: string
      responses:
        '200':
          description: 返回定时任务列表
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
  /sys/schedule/info/{id}:
    get:
      tags:
      - 定时任务
      summary: 获取定时任务信息
      produces:
      - application/json
      parameters:
      - name: jobId
        description: 定时任务ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回定时任务信息
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  SysDeptEntity:
    type: object
    properties:
      deptId:
        description: 部门ID
        type: integer
        format: int64
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID，一级部门为0
        type: integer
        format: int64
      parentName:
        description: 上级部门名称
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysDeptEntityEdit:
    type: object
    properties:
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门ID，一级部门为0
        type: integer
        format: int64
      orderNum:
        description: 排序
        type: integer
        format: int32
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
...
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
  SysLogEntity:
    type: object
    properties:
      id:
        description: 日志ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      operation:
        description: 用户操作
        type: string
      method:
        description: 请求方法
        type: string
      params:
        description: 请求参数
        type: string
      time:
        description: 执行时长(毫秒)
        type: integer
        format: int64
      ip:
        description: IP地址
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysConfigEntity:
    type: object
    properties:
      id:
        description: 参数ID
        type: integer
        format: int64
      key:
        description: 参数名
        type: string
      value:
        description: 参数值
        type: string
      remark:
        description: 备注
        type: string
...
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time